auth/requirements.txt
Flask==2.2.3
gevent==22.10.1
python-dotenv==1.0.0
gunicorn==20.1.0
pydantic[email]==1.10.7
Flask-SQLAlchemy==3.0.3
SQLAlchemy-Utils==0.40.0
psycopg2-binary==2.9.6
redis==4.5.4
alembic==1.10.3
flasgger==0.9.5
Flask-Bcrypt==1.0.1
aioredis==2.0.1
backoff==2.2.1
Flask-JWT-Extended==4.4.4
user-agents==2.2.0
flask-redis==0.4.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-flask==0.38b0
opentelemetry-exporter-jaeger==1.17.0
flask-limiter[redis]==3.3.0
Authlib==1.1.0
requests==2.30.0